requirements.txt
numpy==1.19.0
pandas==1.1.0
torch==1.7.0
torchvision==0.8.1
scikit-learn==0.23.1
scipy==1.5.1
Pillow==5.1.0
matplotlib==2.2.2
tqdm==4.54.0